backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1

python-dotenv==1.0.1
requests==2.32.3

google-generativeai==0.8.3
PyYAML==6.0.2
sentence-transformers==3.3.1
scikit-learn==1.5.2
numpy==2.1.3
